Things mostly working. Added dryrun feature, image rename based on rom title, not filename. Added fanart support. Added license and did a bit of cleanup. Added option to merge images to LB artwork dirs.

+++ lb2am.py
@@ -1,3 +1,20 @@
+# lb2am.py - https://github.com/sharkusk/lb2am
+# Copyright (C) 2017 - Marcus Kellerman
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
 import xml.etree.ElementTree as ET
 import argparse
 import os
@@ -11,10 +28,14 @@ parser.add_argument('AttractMode_dir', help="Base Directory of AttractMode")
 parser.add_argument('--genroms', action="store_true", help="Generate AttractMode romlists from Launchbox data.")
 parser.add_argument('--genplats', action="store_true", help="Generate AttractMode platforms from Launchbox data.")
 parser.add_argument('--renart', action="store_true", help="Rename artwork in Launchbox to be compatible with AttractMode.  Only the first image for each game in each directory will be renamed.")
-parser.add_argument('-r', '--rlauncher', default='', help="Specify RocketLauncher executable, emulators are generated using RocketLauncher settings.")
-parser.add_argument('-e', '--romext', default='.smc;.zip;.7z;.nes;.gba;.gb;.rom;.a26;.lnx;.gg;.int;.sms;.nds;.pce;.cue;.pbp;.iso;.cso;.32x;.bin;.rar;.dsk;.mx2;.lha;.n64;.wud;.wux;.rpx;.cdi;.adf;.d64;.t64', help="Override default rom extention (separated by ';')")
+parser.add_argument('--mergeart', action="store_true", help="Move missing artwork from AttractMode's scraper directory to Launchbox directories.")
+parser.add_argument('--dryrun', action="store_true", help="Don't modify or create any files, only print operations that will be performed.")
+parser.add_argument('--verbose', action="store_true", help="Dump the romlist and platform files to the console.")
+parser.add_argument('--rlauncher', default='', help="Specify RocketLauncher executable, emulators are generated using RocketLauncher settings.")
+parser.add_argument('-e', '--romext', default='.smc;.zip;.7z;.nes;.gba;.gb;.rom;.a26;.lnx;.gg;.int;.sms;.nds;.pce;.cue;.pbp;.iso;.cso;.32x;.bin;.rar;.dsk;.mx2;.lha;.n64;.wud;.wux;.rpx;.cdi;.adf;.d64;.t64',
+        help="Override default rom extention (separated by ';')")
+# TODO Create platform specific rom extensions
 # TODO Specify single platform
-# TODO Specify romlist or emulator import
 # TODO Pick absolute or relative paths to be used
 # TODO Overwrite, update or skip if file exists
 
@@ -41,7 +62,7 @@ AM_FIELD_MAP = [
     { "LB_Field": None,                 "AM_Field": "AltTitle",     "MapToAm": None },
     { "LB_Field": None,                 "AM_Field": "Extra",        "MapToAm": None },
     { "LB_Field": None,                 "AM_Field": "Buttons",      "MapToAm": None }, ]
-        
+
 def ConvertToAMRomlist( LBPlatformFilePath ):
     # import pdb; pdb.set_trace()
 
@@ -58,7 +79,6 @@ def ConvertToAMRomlist( LBPlatformFilePath ):
         # Step through each field needed in the AttractMode emulator file
         for field in AM_FIELD_MAP:
             t = u''
-            # Special case the emulator AttractMode field
             # Check if Launchbox has a field that maps to AttractMode
             if field["LB_Field"]:
                 try:
@@ -72,36 +92,43 @@ def ConvertToAMRomlist( LBPlatformFilePath ):
                 except:
                     pass
             if t:
-                # print(t.encode('utf-8')),
                 output += t
-            # Each entry in the AttractMode is separated by a ;
+            # Each entry in AttractMode is separated by a ;
             output += ';'
-            # print(';'),
         # We are all done with this entry
         output += '\n'
-        # print('')
 
     # Sort the romlist
-    output = '\n'.join(sorted(output.split('\n')))
+    output = '\n'.join(sorted(output.strip().split('\n')))
 
-    # print output.encode('utf-8')
     return output
 
 def CreateRomlists( LaunchboxBaseDir, AttractModeBaseDir ):
     platformsdir = os.path.join(LaunchboxBaseDir, 'Data', 'Platforms')
     romlistsdir = os.path.join(AttractModeBaseDir, 'romlists')
 
+    # LB stores roms in individual xml files named after the platform
     files = glob.glob(os.path.join(platformsdir,"*.xml"))
     for file in files:
-        print("Creating RomList for: "+file)
+        print("Extracting ROMS from: "+file)
+        # We use EmulatorName in a lamba, which needs to be global
         global EmulatorName
         EmulatorName = os.path.splitext(os.path.split(file)[1])[0] # Use platform filename as emulator name
-        with codecs.open(os.path.join(romlistsdir,EmulatorName+'.txt'), 'w', 'utf-8') as fout:
-            fout.write(ConvertToAMRomlist(file))
-            fout.close()
+        # LB may use unicode characters, so encode accordingly
+        romListFileName = os.path.join(romlistsdir,EmulatorName+'.txt')
+        output = ConvertToAMRomlist(file)
+        if args.dryrun or args.verbose:
+            print( ("Creating romlist: "+romListFileName).encode('utf-8') )
+            if args.verbose:
+                print( output.encode('utf-8') )
+                print('')
+        else:
+            with codecs.open( romListFileName, 'w', 'utf-8') as fout:
+                fout.write(output)
+                fout.close()
 
 ATTRACTMODE_EMULATOR_FILE_FORMAT = """#
-# Generated by lb2am.py
+# Generated by lb2am.py - https://github.com/sharkusk/lb2am
 #
 executable           %(appPath)s
 args                 %(commandLine)s
@@ -112,94 +139,141 @@ info_source          thegamesdb.net
 %(artwork)s
 """
 
-AM_IMAGES = { 
-        "artwork flyer     ": [ "Images/%(platformName)s/Box - Front", "Images/%(platformName)s/Advertisement Flyer - Front", "Images/%(platformName)s/Box - 3D", "Images/%(platformName)s/Arcade - Cabinet", ],
-        "artwork marquee   ": [ "Images/%(platformName)s/Banner", "Images/%(platformName)s/Arcade - Marquee", ],
-        "artwork snap      ": [ "Images/%(platformName)s/Screenshot - Gameplay", "Images/%(platformName)s/Screenshot - Game Title", "Images/%(platformName)s/Screenshot - Game Select", "Videos/%(platformName)s", ],
-        "artwork wheel     ": [ "Images/%(platformName)s/Clear Logo", ],
+AM_IMAGES = {
+        "artwork flyer     ": [ os.path.join("Images","%(platformName)s","Box - Front"),
+                                os.path.join("Images","%(platformName)s","Advertisement Flyer - Front"),
+                                os.path.join("Images","%(platformName)s","Box - 3D"),
+                                os.path.join("Images","%(platformName)s","Arcade - Cabinet"), ],
+        "artwork marquee   ": [ os.path.join("Images","%(platformName)s","Banner"),
+                                os.path.join("Images","%(platformName)s","Arcade - Marquee"), ],
+        "artwork snap      ": [ os.path.join("Images","%(platformName)s","Screenshot - Gameplay"),
+                                os.path.join("Images","%(platformName)s","Screenshot - Game Title"),
+                                os.path.join("Images","%(platformName)s","Screenshot - Game Select"),
+                                os.path.join("Videos","%(platformName)s"), ],
+        "artwork wheel     ": [ os.path.join("Images","%(platformName)s","Clear Logo"), ],
+        "artwork fanart    ": [ os.path.join("Images","%(platformName)s","Fanart - Background"), ],
 }
 
-AM_IMAGE_REGIONS = [ "Europe", "Japan", "North America", "United States", ]
+AM_IMAGE_REGIONS = [ "United States", "North America", "Europe", "Japan", ]
 
 def CreateAmEmulators( LaunchboxBaseDir, AttractModeBaseDir ):
     tree = ET.parse(os.path.join(LaunchboxBaseDir, 'Data', 'Emulators.xml'))
     root = tree.getroot()
 
     # Emulators and platforms are separated in LB, they are cross referenced through a unique ID
-    EM_FIELDS = [ 'Title', 'ApplicationPath', 'CommandLine', 'NoSpace', ]
-    emulatordict = {} # { ID: { "Title": title, "ApplicationPath": path, "CommandLine": cmdline } }
+    EM_FIELDS = [ 'Title', 'ApplicationPath', 'CommandLine', 'NoSpace', 'NoQuotes', ]
+    emulatordict = {}
+
+    # Create a dictionary of emulator information we can use when parsing the
+    # LB emulator platforms.  Format is as follows:
+    # { ID: { "Title": title, "ApplicationPath": path, "CommandLine": cmdline } }
     for emulator in root.findall('Emulator'):
         emulatorEntry = {}
         for field in EM_FIELDS:
             emulatorEntry[field] = emulator.find(field).text
         emulatordict[emulator.find('ID').text] = emulatorEntry
 
-
+    # Now we can parse each emulator platform and combine with the information
+    # stored in the dictionary above to create an AM emulator
     for emulatorPlatform in root.findall('EmulatorPlatform'):
         if emulatorPlatform.find('Default').text == 'true':
             platformName = emulatorPlatform.find('Platform').text
             print("Creating Emulator: "+platformName)
+
             if args.rlauncher:
                 appPath = os.path.abspath(args.rlauncher)
                 commandLine = 'args -s "[emulator]" -r "[name]" -p AttractMode -f "%s"' % os.path.join(os.path.abspath(AttractModeBaseDir), 'attract.exe')
+                romPath = ''
             else:
+                # Launchbox stores the rom path for each rom, while Attractmode uses
+                # a list of rompaths for each emulator
+                romPath = []
+                try:
+                    plattree = ET.parse(os.path.join(LaunchboxBaseDir, 'Data', 'Platforms', platformName+'.xml'))
+                    platroot = plattree.getroot()
+                    for game in platroot.findall('Game'):
+                        romPath.append(os.path.abspath(os.path.split(game.find('ApplicationPath').text)[0]))
+                except:
+                    pass
+
+                # Remove duplicates and convert to string
+                romPath = list(set(romPath))
+                romPath = ';'.join(romPath)
+
+                # Lookup the application path for this emulator (using our dictionary)
                 appPath = emulatordict[emulatorPlatform.find('Emulator').text]['ApplicationPath']
                 appPath = os.path.abspath(os.path.join(LaunchboxBaseDir, appPath))
                 commandLine = emulatorPlatform.find('CommandLine').text
                 if not commandLine:
                     commandLine = ''
-                # TODO Map front-end commandline options between LB and AM
-                if emulatordict[emulatorPlatform.find('Emulator').text]['NoSpace'] == 'false': 
+                # TODO Map other front-end commandline options between LB and AM
+                if emulatordict[emulatorPlatform.find('Emulator').text]['NoSpace'] == 'false':
                     commandLine += ' '
-                commandLine += '[romfilename]'
-            # TODO Use Platforms.xml to obtain default ROM path
-            romPath = '.'
+                if emulatordict[emulatorPlatform.find('Emulator').text]['NoQuotes'] == 'true':
+                    commandLine += '[romfilename]'
+                else:
+                    commandLine += '"[romfilename]"'
             romExt = args.romext
 
-            artwork = ''
-            for k in AM_IMAGES.keys():
-                artwork += k
-                for d in AM_IMAGES[k]:
-                    d = d % { 'platformName': platformName }
-                    artwork += os.path.join(os.path.abspath(LaunchboxBaseDir), d)+';'
+            artworkText = ''
+            for artPrefix in AM_IMAGES.keys():
+                artworkText += artPrefix
+                for artDirNames in AM_IMAGES[artPrefix]:
+                    artDirNames = artDirNames % { 'platformName': platformName }
+                    artworkText += os.path.join(os.path.abspath(LaunchboxBaseDir), artDirNames)+';'
                     for region in AM_IMAGE_REGIONS:
-                        artwork += os.path.join(os.path.abspath(LaunchboxBaseDir), d, region)+';'
-                artwork += '\n'
+                        artworkText += os.path.join(os.path.abspath(LaunchboxBaseDir), artDirNames, region)+';'
+                artworkText += '\n'
 
-            # import pdb; pdb.set_trace()
-            output = ATTRACTMODE_EMULATOR_FILE_FORMAT % { "appPath": appPath, "commandLine": commandLine, "romPath": romPath, "romExt": romExt, "platformName": platformName, "artwork": artwork }
+            output = ATTRACTMODE_EMULATOR_FILE_FORMAT % { "appPath": appPath, "commandLine": commandLine, "romPath": romPath, "romExt": romExt, "platformName": platformName, "artwork": artworkText }
 
-            # Attractmode uses Unix style paths, so replace the \'s
-            output = output.replace('\\', '/')
+            # Attractmode uses Unix style paths, so replace the windows \'s
+            output = output.replace('\\', '/').strip()
 
-            with codecs.open(os.path.join(AttractModeBaseDir,'emulators',platformName+'.cfg'), 'w', 'utf-8') as fout:
-                fout.write(output)
-                fout.close()
+            platformFileName = os.path.join(AttractModeBaseDir,'emulators',platformName+'.cfg')
+
+            if args.dryrun or args.verbose:
+                print( ("Writing emulator file: "+platformFileName).encode('utf-8') )
+                if args.verbose:
+                    print( output.encode('utf-8') )
+                    print('')
+            else:
+                with codecs.open( platformFileName, 'w', 'utf-8') as fout:
+                    fout.write(output)
+                    fout.close()
 
 def RenameLBArtwork( LaunchboxBaseDir, AttractModeBaseDir ):
     romlistsdir = os.path.join(AttractModeBaseDir, 'romlists')
 
     files = glob.glob(os.path.join(romlistsdir,"*.txt"))
-    for f in files:
-        platformName = os.path.splitext(os.path.split(f)[1])[0]
+    for romListFileName in files:
+        platformName = os.path.splitext(os.path.split(romListFileName)[1])[0]
         print("Renaming artwork for: " + platformName)
 
+        # Create list of platform specific artwork directories
         artDirs = []
         for key in AM_IMAGES.keys():
             for value in AM_IMAGES[key]:
+                # Insert platform name into each entry
                 value = value % { 'platformName': platformName }
                 artDirs.append(os.path.join(os.path.abspath(LaunchboxBaseDir), value))
+                # LB also stores images in specific region specific folders
                 for region in AM_IMAGE_REGIONS:
                     artDirs.append(os.path.join(os.path.abspath(LaunchboxBaseDir), value, region))
 
-        file = open(f)
-        for line in file:
-            if line.lstrip()[0] == '#':
-                continue
-            try:
-                gameFileName, gameName, __ = line.split(';', 2)
-            except:
-                continue
+        plattree = ET.parse(os.path.join(LaunchboxBaseDir, 'Data', 'Platforms', platformName+'.xml'))
+        platroot = plattree.getroot()
+        for game in platroot.findall('Game'):
+            # LB uses the game title (not filename) for images when scraping.
+            # We need to rename these files to match the rom filename for AM
+            gameFileName = os.path.splitext(os.path.split(game.find('ApplicationPath').text)[1])[0]
+            gameName = game.find('Title').text
+
+            # When LB saves image files, it replaces the following chacters with '_'
+            LB_FILE_SUB = [ ':', "'", '\\', '/', '"', '?', '<', '>', '!', '|' ]
+
+            for sub in LB_FILE_SUB:
+                gameName = gameName.replace( sub, '_' )
 
             images = []
             for artDir in artDirs:
@@ -207,10 +281,40 @@ def RenameLBArtwork( LaunchboxBaseDir, AttractModeBaseDir ):
                 images.extend(glob.glob(searchString))
 
             for image in images:
-                __, ext = os.path.splitext(image)
+                # First, extract the extension
+                ext = os.path.splitext(image)[1]
+                # Remove the old filename (keep the path), then append the new filename and old extension
                 newImage = os.path.join(os.path.split(image)[0],gameFileName+ext)
-                # print( "Renaming "+image+" to "+newImage )
-                os.rename(image,newImage)
+                if args.dryrun:
+                    print( ("Renaming "+image+" to "+newImage).encode('utf-8') )
+                else:
+                    os.rename(image,newImage)
+
+AM_TO_LB_ART_PATH = [
+        ( os.path.join("scraper","%(platformName)s","flyer"),   os.path.join("Images","%(platformName)s","Box - Front") ),
+        ( os.path.join("scraper","%(platformName)s","marquee"), os.path.join("Images","%(platformName)s","Banner") ),
+        ( os.path.join("scraper","%(platformName)s","wheel"),   os.path.join("Images","%(platformName)s","Clear Logo") ),
+        ( os.path.join("scraper","%(platformName)s","fanart"),  os.path.join("Images","%(platformName)s","Fanart - Background") ),
+    ]
+def MergeArtworkToLB( LaunchboxBaseDir, AttractModeBaseDir ):
+    # import pdb; pdb.set_trace()
+    # Get list of directories in AM's scraper directory
+    amScraperPlats = next(os.walk(os.path.join(AttractModeBaseDir, 'scraper')))[1]
+    for platformName in amScraperPlats:
+        for artPath in AM_TO_LB_ART_PATH:
+            searchString = os.path.join(AttractModeBaseDir, artPath[0] % { "platformName": platformName }, '*')
+            imagesToMove = glob.glob(searchString)
+            for image in imagesToMove:
+                # Get the filename and add it to the LB art path
+                newPath = os.path.join(LaunchboxBaseDir, artPath[1] % { "platformName": platformName }, os.path.split(image)[1])
+                # Next, see if it exists in LB
+                if os.path.isfile( newPath ):
+                    # File already exists, so we can move to the next one
+                    continue
+                if args.dryrun:
+                    print( ("Moving "+image+" to "+newPath).encode('utf-8') )
+                else:
+                    shutil.move(image, newPath)
 
 if args.genroms:
     CreateRomlists( args.Launchbox_dir, args.AttractMode_dir )
@@ -218,3 +322,5 @@ if args.genplats:
     CreateAmEmulators( args.Launchbox_dir, args.AttractMode_dir )
 if args.renart:
     RenameLBArtwork( args.Launchbox_dir, args.AttractMode_dir )
+if args.mergeart:
+    MergeArtworkToLB( args.Launchbox_dir, args.AttractMode_dir )
